requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.60.0
numpy>=1.20.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
Pillow>=9.0.0
joblib>=1.1.0
scipy>=1.8.0